cmd/client: drive the real jetstream client; keep raw loadtest mode
The default command now uses the public jetstream client: a live tail, or
a filtered backfill+cutover when --after-seq/--before-seq is given. Flags
map to the client options (collections, dids, seq bounds, live cursor,
batch size, download concurrency, optional --live-buffer-file for the
durable JSONL buffer). --print emits each decoded event as JSON;
otherwise it prints periodic throughput stats.

The legacy raw-websocket load tester is preserved verbatim under the
"loadtest" subcommand (moved to loadtest.go), so the direct
multi-connection /subscribe stress path is still available.

Also add JSON struct tags to the public Event/Commit/Identity/Account/
Sync so json.Marshal(Event) yields the familiar Jetstream wire shape
(cursor/time_us/kind/commit/record/record_cbor/...) — makes --print
output match the websocket payload and the type round-trippable.

Tested: subscribe command drives the real client against a fake
/subscribe-v2 server and prints decoded events; existing loadtest
URL/dial tests retained. -race clean, lint 0.

Closes #89
